requirements.txt
Flask
flask-mysqldb
Flask-JWT-Extended
Flask-Cors
python-dotenv
bcrypt
Flask-Migrate
Flask-SQLAlchemy
Flask-JWT-Extended>=4.6.0
